Split Zellij workflow tests by responsibility
The serialized Zellij workflow coverage combined fleet discovery, presentation and creation, reconnect recovery, and confirmed kill behavior in one 3,700-line source file. Those behaviors have distinct seams, but the shared file made focused ownership and review difficult.

Keep all 58 tests in the original serialized suite while moving each seam to a responsibility-named extension. Shared fixture access expands only within the test target; production code, test identities, and scheduling remain unchanged.

The aggregate suite also depended on whether Ghosthub was installed in Applications because resource-locator fixtures used the real packaged path. Use a unique nonexistent temporary app path so those fallback contracts stay isolated from developer installations.
